refactor : 각 Creator 메서드 및 패키지명 변경
- entityCreator -> entityfactory로 변경
- 클래스명도 00Factory로 변경
- ImageFactory는 Image로 분리
